+++ src/Controller/JugadorController.php
@@ -8,7 +8,9 @@ use App\Repository\JugadorRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Routing\Annotation\Route;
+use Twig\Node\Expression\Test\SameasTest;
 
 /**
  * @Route("/jugador")
@@ -18,9 +20,25 @@ class JugadorController extends AbstractController
     /**
      * @Route("/", name="jugador_index", methods={"GET"})
      */
-    public function index(JugadorRepository $jugadorRepository): Response
+    public function index(JugadorRepository $jugadorRepository, SessionInterface $session): Response
     {
+        if ($session->get('arbitreLogged') ==null){
+            return $this->render('index.html.twig', [
+                'controller_name' => 'MainController',
+            ]);
+        }
         return $this->render('jugador/index.html.twig', [
+            'jugadors' => $jugadorRepository->findAll(),
+        ]);
+    }
+
+    /**
+     * @Route("/index", name="jugador_index_clean", methods={"GET"})
+     */
+    public function indexClean(JugadorRepository $jugadorRepository, SessionInterface $session): Response
+    {
+        
+        return $this->render('jugador/indexClean.html.twig', [
             'jugadors' => $jugadorRepository->findAll(),
         ]);
     }
@@ -28,8 +46,13 @@ class JugadorController extends AbstractController
     /**
      * @Route("/new", name="jugador_new", methods={"GET","POST"})
      */
-    public function new(Request $request): Response
+    public function new(Request $request, SessionInterface $session): Response
     {
+        if ($session->get('arbitreLogged') ==null){
+            return $this->render('index.html.twig', [
+                'controller_name' => 'MainController',
+            ]);
+        }
         $jugador = new Jugador();
         $form = $this->createForm(JugadorType::class, $jugador);
         $form->handleRequest($request);
@@ -51,8 +74,13 @@ class JugadorController extends AbstractController
     /**
      * @Route("/{id}", name="jugador_show", methods={"GET"})
      */
-    public function show(Jugador $jugador): Response
+    public function show(Jugador $jugador, SessionInterface $session): Response
     {
+        if ($session->get('arbitreLogged') ==null){
+            return $this->render('index.html.twig', [
+                'controller_name' => 'MainController',
+            ]);
+        }
         return $this->render('jugador/show.html.twig', [
             'jugador' => $jugador,
         ]);
@@ -61,8 +89,13 @@ class JugadorController extends AbstractController
     /**
      * @Route("/{id}/edit", name="jugador_edit", methods={"GET","POST"})
      */
-    public function edit(Request $request, Jugador $jugador): Response
+    public function edit(Request $request, Jugador $jugador, SessionInterface $session): Response
     {
+        if ($session->get('arbitreLogged') ==null){
+            return $this->render('index.html.twig', [
+                'controller_name' => 'MainController',
+            ]);
+        }
         $form = $this->createForm(JugadorType::class, $jugador);
         $form->handleRequest($request);
 
@@ -81,9 +114,14 @@ class JugadorController extends AbstractController
     /**
      * @Route("/{id}", name="jugador_delete", methods={"DELETE"})
      */
-    public function delete(Request $request, Jugador $jugador): Response
+    public function delete(Request $request, Jugador $jugador, SessionInterface $session): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$jugador->getId(), $request->request->get('_token'))) {
+        if ($session->get('arbitreLogged') ==null){
+            return $this->render('index.html.twig', [
+                'controller_name' => 'MainController',
+            ]);
+        }
+        if ($this->isCsrfTokenValid('delete' . $jugador->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($jugador);
             $entityManager->flush();
